Move upgradeDependencies to /src/lib
