+++ src/versionmanager.ts
@@ -8,18 +8,12 @@ import ProgressBar from 'progress'
 import prompts from 'prompts'
 import { and } from 'fp-and-or'
 import * as versionUtil from './version-util'
-import { FilterPattern, GetVersion, IgnoredUpgrade, Index, Maybe, Options, PackageManager, PackageFile, Version, VersionDeclaration } from './types'
 import getPreferredWildcard from './lib/getPreferredWildcard'
 import isUpgradeable from './lib/isUpgradeable'
 import queryVersions from './lib/queryVersions'
 import getPackageManager from './lib/getPackageManager'
-
-interface MappedDependencies {
-  current: VersionDeclaration,
-  currentParsed: VersionDeclaration | null,
-  latest: Version,
-  latestParsed: Version | null,
-}
+import upgradeDependencies from './lib/upgradeDependencies'
+import { FilterPattern, GetVersion, IgnoredUpgrade, Index, Maybe, Options, PackageManager, PackageFile, Version, VersionDeclaration } from './types'
 
 /**
  * Return true if the version satisfies the range.
@@ -31,74 +25,7 @@ interface MappedDependencies {
  */
 export const isSatisfied = semver.satisfies
 
-/**
- * Upgrade a dependencies collection based on latest available versions. Supports npm aliases.
- *
- * @param currentDependencies current dependencies collection object
- * @param latestVersions latest available versions collection object
- * @param [options={}]
- * @returns upgraded dependency collection object
- */
-function upgradeDependencies(currentDependencies: Index<VersionDeclaration>, latestVersions: Index<Version>, options: Options = {}): Index<VersionDeclaration> {
-  // filter out dependencies with empty values
-  currentDependencies = cint.filterObject(currentDependencies, (key, value) => !!value)
-
-  // get the preferred wildcard and bind it to upgradeDependencyDeclaration
-  const wildcard = getPreferredWildcard(currentDependencies) || versionUtil.DEFAULT_WILDCARD
-
-  /** Upgrades a single dependency. */
-  const upgradeDep = (current: VersionDeclaration, latest: Version) => versionUtil.upgradeDependencyDeclaration(current, latest, {
-    wildcard,
-    removeRange: options.removeRange
-  })
-
-  return _(currentDependencies)
-    // only include packages for which a latest version was fetched
-    .pickBy((current, packageName) => packageName in latestVersions)
-    // unpack npm alias and git urls
-    .mapValues((current: string, packageName: string) => {
-      const latest = latestVersions[packageName]
-      let currentParsed = null
-      let latestParsed = null
-
-      // parse npm alias
-      if (versionUtil.isNpmAlias(current)) {
-        currentParsed = versionUtil.parseNpmAlias(current)![1]
-        latestParsed = versionUtil.parseNpmAlias(latest)![1]
-      }
-
-      // "branch" is also used for tags (refers to everything after the hash character)
-      if (versionUtil.isGithubUrl(current)) {
-
-        const currentTag = versionUtil.getGithubUrlTag(current)!
-        const [currentSemver] = semverutils.parseRange(currentTag)
-        currentParsed = versionUtil.stringify(currentSemver, 'patch')
-
-        const latestTag = versionUtil.getGithubUrlTag(latest)!
-        const [latestSemver] = semverutils.parseRange(latestTag)
-        latestParsed = versionUtil.stringify(latestSemver, 'patch')
-      }
-
-      return { current, currentParsed, latest, latestParsed }
-    })
-    // pick the packages that are upgradeable
-    .pickBy(({ current, currentParsed, latest, latestParsed }: any) =>
-      isUpgradeable(currentParsed || current, latestParsed || latest)
-    )
-    // pack embedded versions: npm aliases and git urls
-    .mapValues(({ current, currentParsed, latest, latestParsed }: MappedDependencies) => {
-      const upgraded = upgradeDep(currentParsed || current, latestParsed || latest)
-      return versionUtil.isNpmAlias(current) ? versionUtil.upgradeNpmAlias(current, upgraded)
-        : versionUtil.isGithubUrl(current) ? versionUtil.upgradeGithubUrl(current, upgraded)
-        : upgraded
-    })
-    // TODO: type
-    .value() as unknown as Index<VersionDeclaration>
-}
-
-/**
- * @returns String safe for use in `new RegExp()`
- */
+/** Returns a string that is safe to use in `new RegExp(...)`. */
 function escapeRegexp(s: string) {
   return s.replace(/[-/\\^$*+?.()|[\]{}]/g, '\\$&') // Thanks Stack Overflow!
 }
@@ -425,7 +352,6 @@ module.exports = {
   getOwnerPerDependency,
 
   // exposed for testing
-  upgradeDependencies,
   getPeerDependenciesFromRegistry,
   getIgnoredUpgrades,
 }
