+++ test/versionmanager.test.js
@@ -270,53 +270,6 @@ describe('versionmanager', () => {
 
   })
 
-  describe('upgradeDependencies', () => {
-
-    it('upgrade simple, non-semver versions', () => {
-      vm.upgradeDependencies({ foo: '1' }, { foo: '2' }).should.eql({ foo: '2' })
-      vm.upgradeDependencies({ foo: '1.0' }, { foo: '1.1' }).should.eql({ foo: '1.1' })
-      vm.upgradeDependencies({ 'ncu-test-simple-tag': 'v1' }, { 'ncu-test-simple-tag': 'v3' }).should.eql({ 'ncu-test-simple-tag': 'v3' })
-    })
-
-    it('upgrade github dependencies', () => {
-      vm.upgradeDependencies({ foo: 'github:foo/bar#v1' }, { foo: 'github:foo/bar#v2' }).should.eql({ foo: 'github:foo/bar#v2' })
-      vm.upgradeDependencies({ foo: 'github:foo/bar#v1.0' }, { foo: 'github:foo/bar#v2.0' }).should.eql({ foo: 'github:foo/bar#v2.0' })
-      vm.upgradeDependencies({ foo: 'github:foo/bar#v1.0.0' }, { foo: 'github:foo/bar#v2.0.0' }).should.eql({ foo: 'github:foo/bar#v2.0.0' })
-    })
-
-    it('upgrade latest versions that already satisfy the specified version', () => {
-      vm.upgradeDependencies({ mongodb: '^1.0.0' }, { mongodb: '1.4.30' }).should.eql({
-        mongodb: '^1.4.30'
-      })
-    })
-
-    it('do not downgrade', () => {
-      vm.upgradeDependencies({ mongodb: '^2.0.7' }, { mongodb: '1.4.30' }).should.eql({})
-    })
-
-    it('use the preferred wildcard when converting <, closed, or mixed ranges', () => {
-      vm.upgradeDependencies({ a: '1.*', mongodb: '<1.0' }, { mongodb: '3.0.0' }).should.eql({ mongodb: '3.*' })
-      vm.upgradeDependencies({ a: '1.x', mongodb: '<1.0' }, { mongodb: '3.0.0' }).should.eql({ mongodb: '3.x' })
-      vm.upgradeDependencies({ a: '~1', mongodb: '<1.0' }, { mongodb: '3.0.0' }).should.eql({ mongodb: '~3.0' })
-      vm.upgradeDependencies({ a: '^1', mongodb: '<1.0' }, { mongodb: '3.0.0' }).should.eql({ mongodb: '^3.0' })
-
-      vm.upgradeDependencies({ a: '1.*', mongodb: '1.0 < 2.0' }, { mongodb: '3.0.0' }).should.eql({ mongodb: '3.*' })
-      vm.upgradeDependencies({ mongodb: '1.0 < 2.*' }, { mongodb: '3.0.0' }).should.eql({ mongodb: '3.*' })
-    })
-
-    it('convert closed ranges to caret (^) when preferred wildcard is unknown', () => {
-      vm.upgradeDependencies({ mongodb: '1.0 < 2.0' }, { mongodb: '3.0.0' }).should.eql({ mongodb: '^3.0' })
-    })
-
-    it('ignore packages with empty values', () => {
-      vm.upgradeDependencies({ mongodb: null }, { mongodb: '1.4.30' })
-        .should.eql({})
-      vm.upgradeDependencies({ mongodb: '' }, { mongodb: '1.4.30' })
-        .should.eql({})
-    })
-
-  })
-
   describe('getInstalledPackages', function () {
     this.timeout(30000)
     it('execute npm ls', () => {
